requirement.txt
python-chess
pandas
numpy
matplotlib
tqdm
